# docs/reference/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "RevealUI REST API",
    "version": "1.0.0",
    "description": "RevealUI Framework REST API - Modern React 19 Framework with Next.js 16",
    "contact": {
      "name": "RevealUI",
      "url": "https://revealui.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    },
    {
      "url": "https://api.revealui.com/api",
      "description": "Production API server"
    }
  ],
...
  "paths": {
...
    "/auth/sign-up": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign up",
        "description": "Create new user account",
        "operationId": "signUp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "editor", "viewer"]
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "name"]
      },
...
    }
  },
...
}